+++ forma-website/src/components/sections/index.ts
@@ -2,6 +2,7 @@ export { default as HeroSection } from "./HeroSection";
 export { default as CredibilityStrip } from "./CredibilityStrip";
 export { default as FeaturesSection } from "./FeaturesSection";
 export { default as BeforeAfterSection } from "./BeforeAfterSection";
+export { default as SocialProofSection } from "./SocialProofSection";
 export { default as PricingSection } from "./PricingSection";
 export { default as FAQSection } from "./FAQSection";
 export { default as NewsletterSection } from "./NewsletterSection";